==> 1_scripts/9_EntrenamientoModelo_rebalanceo.R <==
## Renovacion de bases de datos (agregando nuevos datos y empresas),
## separacion de base nueva en entrenamiento y prueba y actualizacion
## de las betas para el modelo (manteniendo variables del modelo fijas).
#### paqueteria ----
library(tidyverse)
library(magrittr)
library(stringr)
library(xts)
#### Lectura datos nuevos bases ----
#### Renovar bases ----
#### Separacion en entrenamiento y prueba ----
#### Actualizacion de las betas del modelo ----
